requirements.txt
streamlit>=1.31.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openai>=1.12.0
python-dotenv>=1.0.0
urllib3>=2.0.0
anthropic>=0.18.0 